splitting the shoebox controller, was too large to maintain and already has too many dependencies
