+++ kifi-backend/modules/common/app/com/keepit/shoebox/ShoeboxServiceClient.scala
@@ -96,9 +96,6 @@ trait ShoeboxServiceClient extends ServiceClient {
   def isSensitiveURI(uri: String): Future[Boolean]
   def updateURIRestriction(id: Id[NormalizedURI], r: Option[Restriction]): Future[Unit]
   def sendUnreadMessages(threadItems: Seq[ThreadItem], otherParticipants: Set[Id[User]], user: Id[User], title: String, deepLocator: DeepLocator, notificationUpdatedAt: DateTime): Future[Unit]
-  def getAllURLPatterns(): Future[Seq[UrlPatternRule]]
-  def updateScreenshots(nUriId: Id[NormalizedURI]): Future[Unit]
-  def getUriImage(nUriId: Id[NormalizedURI]): Future[Option[String]]
   def getUriSummary(request: URISummaryRequest): Future[URISummary]
   def getURISummaries(uriIds: Seq[Id[NormalizedURI]]): Future[Map[Id[NormalizedURI], URISummary]]
   def getUserImageUrl(userId: Id[User], width: Int): Future[String]
@@ -129,8 +126,7 @@ case class ShoeboxCacheProvider @Inject() (
   userConnCountCache: UserConnectionCountCache,
   userBookmarkCountCache: KeepCountCache,
   userSegmentCache: UserSegmentCache,
-  extensionVersionCache: ExtensionVersionInstallationIdCache,
-  urlPatternRuleAllCache: UrlPatternRuleAllCache)
+  extensionVersionCache: ExtensionVersionInstallationIdCache)
 
 class ShoeboxServiceClientImpl @Inject() (
   override val serviceCluster: ServiceCluster,
@@ -614,24 +610,6 @@ class ShoeboxServiceClientImpl @Inject() (
       "notificationUpdatedAt" -> notificationUpdatedAt
     )
     call(Shoebox.internal.sendUnreadMessages(), payload).imap(_ => {})
-  }
-
-  def getAllURLPatterns(): Future[Seq[UrlPatternRule]] = {
-    cacheProvider.urlPatternRuleAllCache.getOrElseFuture(UrlPatternRuleAllKey()) {
-      call(Shoebox.internal.allURLPatternRules()).map { r =>
-        Json.fromJson[Seq[UrlPatternRule]](r.json).get
-      }
-    }
-  }
-
-  def updateScreenshots(nUriId: Id[NormalizedURI]): Future[Unit] = {
-    call(Shoebox.internal.updateScreenshots(nUriId)).map { r => assert(r.status == 202); () }
-  }
-
-  def getUriImage(nUriId: Id[NormalizedURI]): Future[Option[String]] = {
-    call(Shoebox.internal.getUriImage(nUriId)).map { r =>
-      Json.fromJson[Option[String]](r.json).get
-    }
   }
 
   def getUriSummary(request: URISummaryRequest): Future[URISummary] = {
