+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/internal/ShoeboxController.scala
@@ -1,37 +1,29 @@
 package com.keepit.controllers.internal
 
-import com.google.inject.{ Provider, Inject }
+import com.google.inject.Inject
+import com.keepit.commanders.{ KeepsCommander, RawKeepImporterPlugin, UserCommander }
+import com.keepit.common.akka.SafeFuture
 import com.keepit.common.controller.ShoeboxServiceController
-import com.keepit.common.db.{ ExternalId, Id }
 import com.keepit.common.db.slick.Database
+import com.keepit.common.db.{ ExternalId, Id }
 import com.keepit.common.healthcheck.AirbrakeNotifier
 import com.keepit.common.logging.Logging
 import com.keepit.common.mail.{ ElectronicMail, LocalPostOffice }
 import com.keepit.common.service.FortyTwoServices
 import com.keepit.common.social.BasicUserRepo
 import com.keepit.common.time._
+import com.keepit.heimdal.SanitizedKifiHit
 import com.keepit.model._
 import com.keepit.normalizer._
+import com.keepit.scraper._
 import com.keepit.search.{ SearchConfigExperiment, SearchConfigExperimentRepo }
-import com.keepit.common.performance._
+import com.keepit.social.{ BasicUser, SocialGraphPlugin, SocialId, SocialNetworkType }
 import org.joda.time.DateTime
-
-import scala.concurrent.Future
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
-
 import play.api.libs.json._
 import play.api.mvc.Action
-import com.keepit.scraper._
-import com.keepit.social.{ SocialGraphPlugin, BasicUser, SocialNetworkType }
 
-import com.keepit.commanders.{ KeepsCommander, RawKeepImporterPlugin, UserCommander }
-import com.keepit.normalizer.VerifiedCandidate
-import com.keepit.model.KifiInstallation
-import com.keepit.social.SocialId
-import play.api.libs.json.JsObject
-import scala.util.{ Try, Failure, Success }
-import com.keepit.common.akka.SafeFuture
-import com.keepit.heimdal.SanitizedKifiHit
+import scala.util.{ Failure, Success }
 
 class ShoeboxController @Inject() (
   db: Database,
@@ -40,8 +32,6 @@ class ShoeboxController @Inject() (
   keepRepo: KeepRepo,
   normUriRepo: NormalizedURIRepo,
   normalizedURIInterner: NormalizedURIInterner,
-  normalizationServiceProvider: Provider[NormalizationService],
-  urlPatternRuleRepo: UrlPatternRuleRepo,
   searchConfigExperimentRepo: SearchConfigExperimentRepo,
   probabilisticExperimentGeneratorRepo: ProbabilisticExperimentGeneratorRepo,
   userExperimentRepo: UserExperimentRepo,
@@ -56,19 +46,14 @@ class ShoeboxController @Inject() (
   searchFriendRepo: SearchFriendRepo,
   emailAddressRepo: UserEmailAddressRepo,
   keepsCommander: KeepsCommander,
-  scrapeInfoRepo: ScrapeInfoRepo,
-  imageInfoRepo: ImageInfoRepo,
-  pageInfoRepo: PageInfoRepo,
   friendRequestRepo: FriendRequestRepo,
   userValueRepo: UserValueRepo,
   userCommander: UserCommander,
   kifiInstallationRepo: KifiInstallationRepo,
   socialGraphPlugin: SocialGraphPlugin,
   rawKeepImporterPlugin: RawKeepImporterPlugin,
-  scraperHelper: ScraperCallbackHelper,
   scrapeScheduler: ScrapeScheduler,
-  verifiedEmailUserIdCache: VerifiedEmailUserIdCache,
-  latestKeepUrlCache: LatestKeepUrlCache)(implicit private val clock: Clock,
+  verifiedEmailUserIdCache: VerifiedEmailUserIdCache)(implicit private val clock: Clock,
     private val fortyTwoServices: FortyTwoServices)
     extends ShoeboxServiceController with Logging {
 
@@ -130,162 +115,6 @@ class ShoeboxController @Inject() (
     Ok(Json.toJson(uri))
   }
 
-  def saveNormalizedURI() = SafeAsyncAction(parse.tolerantJson(maxLength = MaxContentLength)) { request =>
-    val ts = System.currentTimeMillis
-    val normalizedUri = request.body.as[NormalizedURI]
-    val saved = scraperHelper.saveNormalizedURI(normalizedUri)
-    log.debug(s"[saveNormalizedURI] time-lapsed:${System.currentTimeMillis - ts} url=(${normalizedUri.url}) result=$saved")
-    Ok(Json.toJson(saved))
-  }
-
-  def updateNormalizedURI(uriId: Id[NormalizedURI]) = SafeAsyncAction(parse.tolerantJson) { request =>
-    val saveResult = Try {
-      // Handle serialization in session to be transactional.
-      val originalNormalizedUri = db.readOnlyMaster { implicit s => normUriRepo.get(uriId) }
-      val originalJson = Json.toJson(originalNormalizedUri).as[JsObject]
-      val newNormalizedUriResult = Json.fromJson[NormalizedURI](originalJson ++ request.body.as[JsObject])
-
-      newNormalizedUriResult.fold({ invalid =>
-        val error = "Could not deserialize NormalizedURI ($uriId) update: $invalid\nOriginal: $originalNormalizedUri\nbody: ${request.body}"
-        airbrake.notify(error)
-        throw new Exception(error)
-      }, { normalizedUri =>
-        scraperHelper.saveNormalizedURI(normalizedUri)
-      })
-    }
-    saveResult match {
-      case Success(res) =>
-        Ok
-      case Failure(ex) =>
-        log.error(s"Could not deserialize NormalizedURI ($uriId) update: $ex\nbody: ${request.body}")
-        airbrake.notify(s"Could not deserialize NormalizedURI ($uriId) update", ex)
-        throw ex
-    }
-  }
-
-  // todo: revisit
-  def recordPermanentRedirect() = Action.async(parse.tolerantJson) { request =>
-    val ts = System.currentTimeMillis
-    log.debug(s"[recordPermanentRedirect] body=${request.body}")
-    val args = request.body.as[JsArray].value
-    require(!args.isEmpty && args.length == 2, "Both uri and redirect need to be supplied")
-    val uri = args(0).as[NormalizedURI]
-    val redirect = args(1).as[HttpRedirect]
-    require(redirect.isPermanent, "HTTP redirect is not permanent.")
-    require(redirect.isLocatedAt(uri.url), "Current Location of HTTP redirect does not match normalized Uri.")
-    val verifiedCandidateOption = normalizationServiceProvider.get.prenormalize(redirect.newDestination).toOption.flatMap { prenormalizedDestination =>
-      db.readWrite { implicit session =>
-        val (candidateUrl, candidateNormalizationOption) = normUriRepo.getByNormalizedUrl(prenormalizedDestination) match {
-          case None => (prenormalizedDestination, SchemeNormalizer.findSchemeNormalization(prenormalizedDestination))
-          case Some(referenceUri) if referenceUri.state != NormalizedURIStates.REDIRECTED => (referenceUri.url, referenceUri.normalization)
-          case Some(reverseRedirectUri) if reverseRedirectUri.redirect == Some(uri.id.get) =>
-            (reverseRedirectUri.url, SchemeNormalizer.findSchemeNormalization(reverseRedirectUri.url))
-          case Some(redirectedUri) =>
-            val referenceUri = normUriRepo.get(redirectedUri.redirect.get)
-            (referenceUri.url, referenceUri.normalization)
-        }
-        candidateNormalizationOption.map(VerifiedCandidate(candidateUrl, _))
-      }
-    }
-
-    val resFutureOption = verifiedCandidateOption.map { verifiedCandidate =>
-      val toBeRedirected = NormalizationReference(uri, correctedNormalization = Some(Normalization.MOVED))
-      val updateFuture = normalizationServiceProvider.get.update(toBeRedirected, verifiedCandidate)
-      // Scraper reports entire NormalizedUri objects with a major chance of stale data / race conditions
-      // The following is meant for synchronisation and should be revisited when scraper apis are rewritten to report modified fields only
-
-      updateFuture.map {
-        case Some(update) => {
-          val redirectedUri = db.readOnlyMaster { implicit session => normUriRepo.get(uri.id.get) }
-          log.debug(s"[recordedPermanentRedirect($uri, $redirect)] time-lapsed: ${System.currentTimeMillis - ts} result=$redirectedUri")
-          redirectedUri
-        }
-        case None => {
-          log.warn(s"[failedToRecordPermanentRedirect($uri, $redirect)] Normalization update failed - time-lapsed: ${System.currentTimeMillis - ts} result=$uri")
-          uri
-        }
-      }
-    }
-
-    val resFuture = resFutureOption getOrElse {
-      log.warn(s"[failedToRecordPermanentRedirect($uri, $redirect)] Redirection normalization empty - time-lapsed: ${System.currentTimeMillis - ts} result=$uri")
-      Future.successful(uri)
-    }
-
-    resFuture.map { res => Ok(Json.toJson(res)) }
-  }
-
-  def recordScrapedNormalization() = Action.async(parse.tolerantJson) { request =>
-    val candidateUrl = (request.body \ "url").as[String]
-    timing(s"recordScrapedNormalization.$candidateUrl") {
-      val candidateNormalization = (request.body \ "normalization").as[Normalization]
-      val scrapedCandidate = ScrapedCandidate(candidateUrl, candidateNormalization)
-
-      val uriId = (request.body \ "id").as[Id[NormalizedURI]](Id.format)
-      val signature = Signature((request.body \ "signature").as[String])
-      val scrapedUri = db.readOnlyMaster { implicit session => normUriRepo.get(uriId) }
-      normalizationServiceProvider.get.update(NormalizationReference(scrapedUri, signature = Some(signature)), scrapedCandidate).map { newReferenceOption =>
-        (request.body \ "alternateUrls").asOpt[Set[String]].foreach { alternateUrls =>
-          val bestReference = newReferenceOption.map { newReferenceId =>
-            db.readOnlyMaster { implicit session =>
-              normUriRepo.get(newReferenceId)
-            }
-          } getOrElse scrapedUri
-          // todo(Léo): What follows is dangerous. Someone could mess up with our data by reporting wrong alternate Urls on its website. We need to do a specific content check.
-          bestReference.normalization.map(ScrapedCandidate(scrapedUri.url, _)).foreach { bestCandidate =>
-            alternateUrls.foreach { alternateUrl =>
-              val uri = db.readOnlyMaster { implicit session =>
-                normalizedURIInterner.getByUri(alternateUrl)
-              }
-              uri match {
-                case Some(existingUri) if existingUri.id.get == bestReference.id.get => // ignore
-                case _ => try {
-                  db.readWrite { implicit session =>
-                    normalizedURIInterner.internByUri(alternateUrl, bestCandidate)
-                  }
-                } catch { case ex: Throwable => log.error(s"Failed to intern alternate url $alternateUrl for $bestCandidate") }
-              }
-            }
-          }
-        }
-        Ok
-      }
-    }
-  }
-
-  def getProxy(url: String) = SafeAsyncAction { request =>
-    val httpProxyOpt = db.readOnlyReplica(2) { implicit session =>
-      urlPatternRuleRepo.getProxy(url)
-    }
-    log.debug(s"[getProxy($url): result=$httpProxyOpt")
-    Ok(Json.toJson(httpProxyOpt))
-  }
-
-  def getProxyP = SafeAsyncAction(parse.tolerantJson) { request =>
-    val url = request.body.as[String]
-    val httpProxyOpt = db.readOnlyReplica(2) { implicit session =>
-      urlPatternRuleRepo.getProxy(url)
-    }
-    Ok(Json.toJson(httpProxyOpt))
-  }
-
-  def isUnscrapable(url: String, destinationUrl: Option[String]) = SafeAsyncAction { request =>
-    val res = urlPatternRuleRepo.rules().isUnscrapable(url) || (destinationUrl.isDefined && urlPatternRuleRepo.rules().isUnscrapable(destinationUrl.get))
-    log.debug(s"[isUnscrapable($url, $destinationUrl)] result=$res")
-    Ok(JsBoolean(res))
-  }
-
-  def isUnscrapableP() = SafeAsyncAction(parse.tolerantJson(maxLength = MaxContentLength)) { request =>
-    val ts = System.currentTimeMillis
-    val args = request.body.as[JsArray].value
-    require(args != null && args.length >= 1, "Expect args to be url && opt[dstUrl] ")
-    val url = args(0).as[String]
-    val destinationUrl = if (args.length > 1) args(1).asOpt[String] else None
-    val res = urlPatternRuleRepo.rules().isUnscrapable(url) || (destinationUrl.isDefined && urlPatternRuleRepo.rules().isUnscrapable(destinationUrl.get))
-    log.debug(s"[isUnscrapableP] time-lapsed:${System.currentTimeMillis - ts} url=$url dstUrl=${destinationUrl.getOrElse("")} result=$res")
-    Ok(JsBoolean(res))
-  }
-
   def getNormalizedURIs(ids: String) = SafeAsyncAction { request =>
     val uriIds = ids.split(',').map(id => Id[NormalizedURI](id.toLong))
     val uris = db.readOnlyMaster { implicit s => uriIds map normUriRepo.get } //using cache
@@ -315,10 +144,9 @@ class ShoeboxController @Inject() (
       normalizedURIInterner.getByUriOrPrenormalize(url) match {
         case Success(Right(prenormalizedUrl)) => Json.obj("url" -> prenormalizedUrl)
         case Success(Left(nuri)) => Json.obj("normalizedURI" -> nuri)
-        case Failure(ex) => {
+        case Failure(ex) =>
           log.error("Could not get normalized uri or prenormalized url", ex)
           Json.obj("url" -> url)
-        }
       }
     }
     Ok(normalizedUriOrPrenormStr)
@@ -335,47 +163,6 @@ class ShoeboxController @Inject() (
     Ok(Json.toJson(uri))
   }
 
-  def assignScrapeTasks(zkId: Id[ScraperWorker], max: Int) = SafeAsyncAction { request =>
-    val res = scraperHelper.assignTasks(zkId, max)
-    Ok(Json.toJson(res))
-  }
-
-  def getImageInfo(id: Id[ImageInfo]) = SafeAsyncAction { request =>
-    val imageInfo = db.readOnlyReplica { implicit ro =>
-      imageInfoRepo.get(id)
-    }
-    log.debug(s"[getImageInfo($id)] result=$imageInfo")
-    Ok(Json.toJson(imageInfo))
-  }
-
-  def saveImageInfo() = SafeAsyncAction(parse.tolerantJson) { request =>
-    val json = request.body
-    val info = json.as[ImageInfo]
-    val saved = scraperHelper.saveImageInfo(info)
-    log.debug(s"[saveImageInfo] result=$saved")
-    Ok(Json.toJson(saved))
-  }
-
-  def savePageInfo() = SafeAsyncAction(parse.tolerantJson) { request =>
-    val json = request.body
-    val info = json.as[PageInfo]
-    val toSave = db.readOnlyMaster { implicit ro => pageInfoRepo.getByUri(info.uriId) } map { p => info.withId(p.id.get) } getOrElse info
-    val saved = scraperHelper.savePageInfo(toSave)
-    log.debug(s"[savePageInfo] result=$saved")
-    Ok
-  }
-
-  def saveScrapeInfo() = SafeAsyncAction(parse.tolerantJson) { request =>
-    val ts = System.currentTimeMillis
-    val json = request.body
-    val info = json.as[ScrapeInfo]
-    val saved = db.readWrite(attempts = 3) { implicit s =>
-      scrapeInfoRepo.save(info)
-    }
-    log.debug(s"[saveScrapeInfo] time-lapsed:${System.currentTimeMillis - ts} result=$saved")
-    Ok
-  }
-
   def getBookmarks(userId: Id[User]) = Action { request =>
     val bookmarks = db.readOnlyReplica(2) { implicit session => //no cache used
       keepRepo.getByUser(userId)
@@ -388,28 +175,6 @@ class ShoeboxController @Inject() (
       keepRepo.getByUriAndUser(uriId, userId)
     }.map(Json.toJson(_)).getOrElse(JsNull)
     Ok(bookmark)
-  }
-
-  def getBookmarksByUriWithoutTitle(uriId: Id[NormalizedURI]) = Action { request =>
-    val ts = System.currentTimeMillis
-    val bookmarks = db.readOnlyReplica(2) { implicit session =>
-      keepRepo.getByUriWithoutTitle(uriId)
-    }
-    log.debug(s"[getBookmarksByUriWithoutTitle($uriId)] time-lapsed:${System.currentTimeMillis - ts} bookmarks(len=${bookmarks.length}):${bookmarks.mkString}")
-    Ok(Json.toJson(bookmarks))
-  }
-
-  def getLatestKeep() = Action(parse.json) { request =>
-    val url = request.body.as[String]
-    val bookmarkOpt = db.readOnlyMaster(2) { implicit session =>
-      latestKeepUrlCache.getOrElseOpt(LatestKeepUrlKey(url)) {
-        normUriRepo.getByNormalizedUrl(url).flatMap { uri =>
-          keepRepo.latestKeep(uri.id.get, url)
-        }
-      }
-    }
-    log.debug(s"[getLatestKeep($url)] $bookmarkOpt")
-    Ok(Json.toJson(bookmarkOpt))
   }
 
   def saveBookmark() = Action(parse.tolerantJson) { request =>
@@ -621,11 +386,6 @@ class ShoeboxController @Inject() (
       normUriRepo.updateURIRestriction(uriId, r)
     }
     Ok
-  }
-
-  def getAllURLPatternRules() = Action { request =>
-    val patterns = urlPatternRuleRepo.rules.rules
-    Ok(Json.toJson(patterns))
   }
 
   def getUserImageUrl(id: Long, width: Int) = Action.async { request =>
